move reconcileElements from Portal to CollabWrapper
